+++ tests/src/env.rs
@@ -6,7 +6,8 @@ mod env_tests {
     use rstest::*;
 
     use crate::utils::{
-        application::env::EnvApp, kube_service::KubeService, run_exec_with_target, service,
+        application::env::EnvApp, kube_service::KubeService, run_command::run_exec_with_target,
+        service,
     };
 
     #[cfg_attr(not(feature = "job"), ignore)]
