+++ tests/src/utils/application.rs
@@ -1,4 +1,5 @@
-use super::{run_exec_targetless, run_exec_with_target, TestProcess};
+use super::TestProcess;
+use crate::utils::run_command::{run_exec_targetless, run_exec_with_target};
 
 pub(crate) mod env;
 pub(crate) mod file_ops;
